+++ src/class/methods.rs
@@ -2,7 +2,7 @@
 
 use crate::{ffi, PyObject, Python};
 use libc::c_int;
-use std::ffi::CString;
+use std::ffi::CStr;
 use std::fmt;
 
 /// `PyMethodDefType` represents different types of Python callable objects.
@@ -35,32 +35,32 @@ pub enum PyMethodType {
     PyInitFunc(ffi::initproc),
 }
 
-#[derive(Copy, Clone, Debug)]
+#[derive(Clone, Debug)]
 pub struct PyMethodDef {
-    pub ml_name: &'static str,
-    pub ml_meth: PyMethodType,
-    pub ml_flags: c_int,
-    pub ml_doc: &'static str,
+    pub(crate) ml_name: &'static CStr,
+    pub(crate) ml_meth: PyMethodType,
+    pub(crate) ml_flags: c_int,
+    pub(crate) ml_doc: &'static CStr,
 }
 
 #[derive(Copy, Clone)]
 pub struct PyClassAttributeDef {
-    pub name: &'static str,
-    pub meth: for<'p> fn(Python<'p>) -> PyObject,
+    pub(crate) name: &'static CStr,
+    pub(crate) meth: for<'p> fn(Python<'p>) -> PyObject,
 }
 
-#[derive(Copy, Clone, Debug)]
+#[derive(Clone, Debug)]
 pub struct PyGetterDef {
-    pub name: &'static str,
-    pub meth: ffi::getter,
-    pub doc: &'static str,
+    pub(crate) name: &'static CStr,
+    pub(crate) meth: ffi::getter,
+    doc: &'static CStr,
 }
 
-#[derive(Copy, Clone, Debug)]
+#[derive(Clone, Debug)]
 pub struct PySetterDef {
-    pub name: &'static str,
-    pub meth: ffi::setter,
-    pub doc: &'static str,
+    pub(crate) name: &'static CStr,
+    pub(crate) meth: ffi::setter,
+    doc: &'static CStr,
 }
 
 unsafe impl Sync for PyMethodDef {}
@@ -73,7 +73,67 @@ unsafe impl Sync for PySetterDef {}
 
 unsafe impl Sync for ffi::PyGetSetDef {}
 
+fn get_name(name: &str) -> &CStr {
+    CStr::from_bytes_with_nul(name.as_bytes())
+        .expect("Method name must be terminated with NULL byte")
+}
+
+fn get_doc(doc: &str) -> &CStr {
+    CStr::from_bytes_with_nul(doc.as_bytes()).expect("Document must be terminated with NULL byte")
+}
+
 impl PyMethodDef {
+    pub(crate) fn get_new_func(&self) -> Option<ffi::newfunc> {
+        if let PyMethodType::PyNewFunc(new_func) = self.ml_meth {
+            Some(new_func)
+        } else {
+            None
+        }
+    }
+
+    pub(crate) fn get_cfunction_with_keywords(&self) -> Option<ffi::PyCFunctionWithKeywords> {
+        if let PyMethodType::PyCFunctionWithKeywords(func) = self.ml_meth {
+            Some(func)
+        } else {
+            None
+        }
+    }
+
+    /// Define a function with no `*args` and `**kwargs`.
+    pub fn cfunction(name: &'static str, cfunction: ffi::PyCFunction, doc: &'static str) -> Self {
+        Self {
+            ml_name: get_name(name),
+            ml_meth: PyMethodType::PyCFunction(cfunction),
+            ml_flags: ffi::METH_NOARGS,
+            ml_doc: get_doc(doc),
+        }
+    }
+
+    /// Define a `__new__` function.
+    pub fn new_func(name: &'static str, newfunc: ffi::newfunc, doc: &'static str) -> Self {
+        Self {
+            ml_name: get_name(name),
+            ml_meth: PyMethodType::PyNewFunc(newfunc),
+            ml_flags: ffi::METH_VARARGS | ffi::METH_KEYWORDS,
+            ml_doc: get_doc(doc),
+        }
+    }
+
+    /// Define a function that can take `*args` and `**kwargs`.
+    pub fn cfunction_with_keywords(
+        name: &'static str,
+        cfunction: ffi::PyCFunctionWithKeywords,
+        flags: c_int,
+        doc: &'static str,
+    ) -> Self {
+        Self {
+            ml_name: get_name(name),
+            ml_meth: PyMethodType::PyCFunctionWithKeywords(cfunction),
+            ml_flags: flags | ffi::METH_VARARGS | ffi::METH_KEYWORDS,
+            ml_doc: get_doc(doc),
+        }
+    }
+
     /// Convert `PyMethodDef` to Python method definition struct `ffi::PyMethodDef`
     pub fn as_method_def(&self) -> ffi::PyMethodDef {
         let meth = match self.ml_meth {
@@ -84,12 +144,20 @@ impl PyMethodDef {
         };
 
         ffi::PyMethodDef {
-            ml_name: CString::new(self.ml_name)
-                .expect("Method name must not contain NULL byte")
-                .into_raw(),
+            ml_name: self.ml_name.as_ptr(),
             ml_meth: Some(meth),
             ml_flags: self.ml_flags,
-            ml_doc: self.ml_doc.as_ptr() as *const _,
+            ml_doc: self.ml_doc.as_ptr(),
+        }
+    }
+}
+
+impl PyClassAttributeDef {
+    /// Define a class attribute.
+    pub fn new(name: &'static str, meth: for<'p> fn(Python<'p>) -> PyObject) -> Self {
+        Self {
+            name: get_name(name),
+            meth,
         }
     }
 }
@@ -105,30 +173,44 @@ impl fmt::Debug for PyClassAttributeDef {
 }
 
 impl PyGetterDef {
+    /// Define a getter.
+    pub fn new(name: &'static str, getter: ffi::getter, doc: &'static str) -> Self {
+        Self {
+            name: get_name(name),
+            meth: getter,
+            doc: get_doc(doc),
+        }
+    }
+
     /// Copy descriptor information to `ffi::PyGetSetDef`
     pub fn copy_to(&self, dst: &mut ffi::PyGetSetDef) {
         if dst.name.is_null() {
-            dst.name = CString::new(self.name)
-                .expect("Method name must not contain NULL byte")
-                .into_raw();
+            dst.name = self.name.as_ptr() as _;
         }
         if dst.doc.is_null() {
-            dst.doc = self.doc.as_ptr() as *mut libc::c_char;
+            dst.doc = self.doc.as_ptr() as _;
         }
         dst.get = Some(self.meth);
     }
 }
 
 impl PySetterDef {
+    /// Define a setter.
+    pub fn new(name: &'static str, setter: ffi::setter, doc: &'static str) -> Self {
+        Self {
+            name: get_name(name),
+            meth: setter,
+            doc: get_doc(doc),
+        }
+    }
+
     /// Copy descriptor information to `ffi::PyGetSetDef`
     pub fn copy_to(&self, dst: &mut ffi::PyGetSetDef) {
         if dst.name.is_null() {
-            dst.name = CString::new(self.name)
-                .expect("Method name must not contain NULL byte")
-                .into_raw();
+            dst.name = self.name.as_ptr() as _;
         }
         if dst.doc.is_null() {
-            dst.doc = self.doc.as_ptr() as *mut libc::c_char;
+            dst.doc = self.doc.as_ptr() as _;
         }
         dst.set = Some(self.meth);
     }
@@ -150,10 +232,10 @@ pub trait PyMethods {
 #[cfg(feature = "macros")]
 pub trait PyMethodsInventory: inventory::Collect {
     /// Create a new instance
-    fn new(methods: &'static [PyMethodDefType]) -> Self;
+    fn new(methods: Vec<PyMethodDefType>) -> Self;
 
     /// Returns the methods for a single `#[pymethods] impl` block
-    fn get(&self) -> &'static [PyMethodDefType];
+    fn get(&'static self) -> &'static [PyMethodDefType];
 }
 
 /// Implemented for `#[pyclass]` in our proc macro code.